refactor: move ConfirmDialog to shared/components and add department delete confirmation

+++ angular/06-hr-portal/src/app/pages/department-page/department-page.ts
@@ -3,6 +3,8 @@ import { DepartmentService } from '../../core/services/department.service';
 import { DepartmentList } from './components/department-list/department-list';
 import { MatButtonModule } from '@angular/material/button';
 import { RouterLink } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialog } from '../../shared/components/confirm-dialog/confirm-dialog';
 
 @Component({
   selector: 'app-department-page',
@@ -12,9 +14,27 @@ import { RouterLink } from '@angular/router';
 })
 export class DepartmentPage {
   private departmentService = inject(DepartmentService);
+  private dialog = inject(MatDialog);
   departments = this.departmentService.departments;
 
   onDelete(id: number) {
-    this.departmentService.deleteDepartment(id);
+    const dialogRef = this.dialog.open(ConfirmDialog, {
+      width: '500px',
+      autoFocus: false,
+      data: {
+        title: 'Delete Department',
+        message: 'Are you sure you want to delete this department?',
+        cancelLabel: 'Cancel',
+        confirmLabel: 'Delete',
+      },
+    });
+
+    dialogRef.afterClosed().subscribe({
+      next: (confirmed) => {
+        if (confirmed) {
+          this.departmentService.deleteDepartment(id);
+        }
+      },
+    });
   }
 }
